backend/requirements.txt
Django==4.2.4
django-ninja==0.22.2
django-cors-headers==4.2.0
google-cloud-storage==2.10.0
google-cloud-container==2.33.0
gunicorn==21.2.0
cryptography==41.0.0
psycopg2-binary==2.9.6
PyJWT==2.2.0
cryptography==41.0.0
python-dotenv==0.21.0
numpy==1.23.5
Pillow==10.0.0
celery==5.3.4
redis==4.5.5
watchdog
django-crontab==0.7.1